Restructure folders for vite
