+++ index.html
@@ -18,7 +18,7 @@
     <link rel="stylesheet" href="css/user_settings_dialog.css">
     <link rel="stylesheet" href="node_modules/leaflet/dist/leaflet.css">
 
-    <script src="modernizr-2.6.2-respond-1.1.0.min.js"></script>
+    <script src="/js/modernizr-2.6.2-respond-1.1.0.min.js"></script>
     <script src="node_modules/leaflet/dist/leaflet.js"></script>
     <script src="node_modules/leaflet-marker-rotation/src/rotatedMarker.js" type="module"></script>
     <script src="node_modules/Leaflet.MultiOptionsPolyline/Leaflet.MultiOptionsPolyline.js"></script>
@@ -3203,21 +3203,20 @@
     </div>
     <!-- Scripts -->
     <script src="node_modules/html2canvas/dist/html2canvas.min.js"></script>
-    <script src="js/screenshot.js"></script>
-    <script src="js/complex.js"></script>
-    <script src="js/real.js"></script>
-    <script src="/jquery-1.11.3.min.js"></script>
-    <script src="/jquery-ui-1.11.4.min.js"></script>
+    <script src="/js/complex.js"></script>
+    <script src="/js/real.js"></script>
+    <script src="/js/jquery-1.11.3.min.js"></script>
+    <script src="/js/jquery-ui-1.11.4.min.js"></script>
     <script src="node_modules/bootstrap/dist/js/bootstrap.min.js"></script>
-    <script src="/jquery.ba-throttle-debounce.js"></script>
-    <script src="/three.min.js"></script>
-    <script src="/FileSaver.js"></script>
-    <script src="/semver.js"></script>
-    <script src="/webm-writer/ArrayBufferDataStream.js"></script>
-    <script src="/webm-writer/BlobBuffer.js"></script>
-    <script src="/webm-writer/WebMWriter.js"></script>
+    <script src="/js/jquery.ba-throttle-debounce.js"></script>
+    <script src="/js/three.min.js"></script>
+    <script src="/js/FileSaver.js"></script>
+    <script src="/js/semver.js"></script>
+    <script src="/js/webm-writer/ArrayBufferDataStream.js"></script>
+    <script src="/js/webm-writer/BlobBuffer.js"></script>
+    <script src="/js/webm-writer/WebMWriter.js"></script>
     <script src="node_modules/lodash/lodash.min.js"></script>
-    <script type="module" src="js/main.js"></script>
+    <script type="module" src="src/main.js"></script>
 
     <dialog class="dialogUpdate">
         <h3>Notice</h4>
