+++ src/flightlog_fields_presenter.js
