+++ src/flightlog_video_renderer.js
