+++ src/graph_config_dialog.js
