+++ src/graph_legend.js
