+++ src/graph_spectrum.js
