+++ src/graph_spectrum_calc.js
