+++ src/graph_spectrum_plot.js
