+++ src/header_dialog.js
