+++ src/keys_dialog.js
