+++ src/laptimer.js
