+++ src/seekbar.js
