+++ src/simple-stats.js
