+++ src/user_settings_dialog.js
